+++ src/app/api/lyrics/route.ts
@@ -1,4 +1,4 @@
-import { NextResponse } from "next/server";
+import { NextRequest, NextResponse } from "next/server";
 import { connectMongoDB } from "@/lib/mongodb";
 import { Artist, Lyrics } from "@/models/model";
 
@@ -25,17 +25,27 @@ export async function GET() {
   return NextResponse.json(lyrics);
 }
 
-export async function DELETE(req: {
-  nextUrl: { searchParams: { get: (arg0: string) => any } };
-}) {
-  const id = req.nextUrl.searchParams.get("id");
+export async function DELETE(req: NextRequest) {
+  try {
+    const url = new URL(req.url);
+    const id = url.searchParams.get("id");
 
-  await connectMongoDB();
-  await Lyrics.findByIdAndDelete(id);
-  return NextResponse.json(
-    { message: "Lyrics deleted successfully" },
-    {
-      status: 200,
+    if (!id) {
+      return NextResponse.json({ error: "ID is required" }, { status: 400 });
     }
-  );
+
+    await connectMongoDB();
+    await Lyrics.findByIdAndDelete(id);
+    return NextResponse.json(
+      { message: "Lyrics deleted successfully" },
+      {
+        status: 200,
+      }
+    );
+  } catch (error) {
+    return NextResponse.json(
+      { error: "Failed to delete lyrics" },
+      { status: 500 }
+    );
+  }
 }
